% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_parenthized}
\alias{is_parenthized}
\title{Is Parenthized}
\usage{
is_parenthized(x)
}
\arguments{
\item{x}{a string, intended to be a `pattern` argument for regexp function}
}
\value{
boolean
}
\description{
Determines if the input argument is surrounded by regexp grouping parentheses.
If the parathenses are imbalanced it throws an error.
}
