% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazmat.R
\name{screen_folder}
\alias{screen_folder}
\title{Screen All R Files in Folder for Hazardous Code}
\usage{
screen_folder(path = ".", quiet = FALSE, recurse = FALSE)
}
\arguments{
\item{path}{a string, the path to the file to be screened.}

\item{quiet}{a boolean, if TRUE then console output is suppressed. Defaults
FALSE.}

\item{recurse}{a boolean, if TRUE then sub-directories of `path` are also
checked for R and Rmd files. Defaults to FALSE.}
}
\value{
`screen_folder` is typically called for it's side effects of printing
information about hazardous material to the console. A tibble of the data
used to create this message is returned invisibly to enable piping.
}
\description{
`screen_folder` applies a set of defined predefined searches (regexps) to all
of the R code in all of the R and Rmd files contained in a folder.
}
